requirements.txt
grpcio
numpy
onnxruntime
vald-client-python
